# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "PHP-NFe API",
        "description": "API RESTful for NFe issuance and cancellation using Hexagonal Architecture & NFePHP",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Local Development Server"
        }
    ],
    "paths": {
...
        "/api/nfe/cancelar": {
            "post": {
                "tags": [
                    "NFe"
                ],
                "summary": "Cancelar NFe",
                "description": "Transmits cancellation event for an authorized NFe to SEFAZ",
                "operationId": "d71321fbc7a95c31a5ea2b87c8d0c764",
                "requestBody": {
                    "description": "Cancellation payload with generic example data",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "id_lote",
                                    "c_orgao",
                                    "cnpj",
                                    "chave_nfe",
                                    "data_hora_evento",
                                    "numero_protocolo",
                                    "justificativa"
                                ],
                                "properties": {
                                    "chave_nfe": {
                                        "type": "string",
                                        "example": "35240800000000000000550010000001011000001010"
                                    },
                                    "numero_protocolo": {
                                        "type": "string",
                                        "example": "135240001234567"
                                    },
                                    "justificativa": {
                                        "type": "string",
                                        "example": "Cancelamento solicitado pelo cliente devido a erro no pedido"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "NFe cancelled successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "sucesso": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "xml_path": {
                                            "type": "string",
                                            "example": "nfe/cancelamento/evento.xml"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
...
}